Entregables/luis_alberto_gonzalez_mendez/APPIUM/behave/Behave/utils/dictionaries/login_screen_text.py
import os
from dotenv import load_dotenv
load_dotenv()

username = os.getenv("USR")
password = os.getenv("PSW")

HOME_TEXTS = {
    'HOME_PAGE': "PRODUCTOS"
}

USUARIOS = {
             "USERNAME": username,
             "PASSWORD": password
        }