Entregables/luis_alberto_gonzalez_mendez/APPIUM/behave/requirements.txt
behave==1.2.6
Appium-Python-Client==2.8.1
python-dotenv~=1.0.0
selenium==4.8.2
